YT-19415: Codestyle
Правки к https://a.yandex-team.ru/review/4179186/details

+++ yt/yt/ytlib/query_tracker_client/config.h
@@ -30,9 +30,9 @@ class TQueryTrackerConnectionConfig
 public:
     THashMap<TString, TQueryTrackerStageConfigPtr> Stages;
 
-    int MaxQueryFileCount;
-    int MaxQueryFileNameSizeBytes;
-    int MaxQueryFileContentSizeBytes;
+    i64 MaxQueryFileCount;
+    i64 MaxQueryFileNameSizeBytes;
+    i64 MaxQueryFileContentSizeBytes;
 
     REGISTER_YSON_STRUCT(TQueryTrackerConnectionConfig);
